test(workflows): share the controlled definition service stub and cover the output dialog; align the input dialog validator setup
