+++ src/modules/Elsa.Studio.Workflows/Components/WorkflowDefinitionEditor/Components/WorkflowProperties/Tabs/InputOutput/Components/Inputs/EditInputDialog.razor.cs
@@ -33,21 +33,17 @@ public partial class EditInputDialog
     [Inject] private IVariableTypeService VariableTypeService { get; set; } = default!;
 
     /// <inheritdoc />
-    protected override async Task OnInitializedAsync()
+    protected override async Task OnParametersSetAsync()
     {
+        // Instantiate the edit context first so that it is available when rendering (which happens as soon as we call an async method on the next line).
         _editContext = new(_model);
+        _validator = new(WorkflowDefinition, Localizer);
+
         _storageDriverDescriptors = (await StorageDriverService.GetStorageDriversAsync()).ToList();
         _variableTypes = (await VariableTypeService.GetVariableTypesAsync()).ToList();
         _uiHints = (await GetUIHintsAsync()).ToList();
         _groupedVariableTypes = _variableTypes.GroupBy(x => x.Category).ToList();
-    }
 
-    /// <inheritdoc />
-    protected override void OnParametersSet()
-    {
-        _editContext = new(_model);
-        _validator = new(WorkflowDefinition, Localizer);
-        
         if (Input == null)
         {
             _model.Name = GetNewInputName(WorkflowDefinition.Inputs);
